# Optimize memory usage in library