--- a/lib/src/index.ts
+++ b/lib/src/index.ts
@@ -1,4 +1,4 @@
-export { Report, ScoredPairs, Occurrence, EncodedSemanticResult, DecodedSemanticResult } from "./lib/analyze/report";
+export { Report, ScoredPairs, Occurrence } from "./lib/analyze/report";
 export { Fragment } from "./lib/analyze/fragment";
 export { Region } from "./lib/util/region";
 export { Pair } from "./lib/analyze/pair";
@@ -7,7 +7,7 @@ export { File } from "./lib/file/file";
 export { TokenizedFile } from "./lib/file/tokenizedFile";
 export { Index } from "./lib/analyze/index";
 export { Dolos } from "./dolos";
-export { SemanticAnalyzer, PairedSemanticGroups, NodeStats, UnpairedSemanticGroups }
+export { SemanticAnalyzer, PairedSemanticGroups, NodeStats, UnpairedSemanticGroups, EncodedSemanticResult, DecodedSemanticResult }
   from "./lib/analyze/SemanticAnalyzer";
 export { SharedFingerprint } from "./lib/analyze/sharedFingerprint";
 export { PairedOccurrence } from "./lib/analyze/pairedOccurrence";