--- a/lib/src/index.ts
+++ b/lib/src/index.ts
@@ -1,11 +1,11 @@
-export { Report, ScoredPairs, Occurrence } from "./lib/analyze/report";
+export { Report } from "./lib/analyze/report";
 export { Fragment } from "./lib/analyze/fragment";
 export { Region } from "./lib/util/region";
 export { Pair } from "./lib/analyze/pair";
 export { Options, CustomOptions } from "./lib/util/options";
 export { File } from "./lib/file/file";
 export { TokenizedFile } from "./lib/file/tokenizedFile";
-export { Index } from "./lib/analyze/index";
+export { Index, Occurrence } from "./lib/analyze/index";
 export { Dolos } from "./dolos";
 export { SemanticAnalyzer, PairedSemanticGroups, NodeStats, UnpairedSemanticGroups, EncodedSemanticResult, DecodedSemanticResult }
   from "./lib/analyze/SemanticAnalyzer";