… lib/src/lib/analyze/pair.ts
… lib/src/lib/analyze/report.ts
… lib/src/lib/file/tokenizedFile.ts
… lib/src/lib/tokenizer/codeTokenizer.ts
… lib/src/lib/util/defaultMap.ts
… lib/src/lib/util/region.ts
… lib/src/lib/util/utils.ts
… lib/src/test/dolos.test.ts
… lib/src/test/pair.test.ts
… lib/src/test/semantic.test.ts
… web/src/api/workers/data.worker.ts
… web/src/types/imports-components.d.ts