With this PR we drastically improve memory usage.

We now avoid all uses of the splat operator (`...`) because it puts all the array items on the stack. causing `Maximum call stack size exceeded` errors.

We also don't calculate fragments anymore for Pairs when doing a regular analysis in the library. This step is slow and uses quadratic (O(n²)) memory. Instead, we calculate metrics (similarity, total and longest overlap) alternatively and only store the resulting metrics. However, the fragments can still be calculated since they are needed for the `--compare` feature of the CLI.

Below we give a bit more details about these improvements.  

# Breaking changes

The API of the lib has a few breaking changes which should be mentioned:
- `Index#createMatches` now returns a `Map<Hash, SharedFingerprint>` 
- The `Report` class now requires a `Map<Hash, SharedFingerprint>` for its constructor. Additional methods like `finish()` have been removed as it is no longer needed to finalize the report.
- The `ScoredPairs` class is removed, all scores/metrics are now present in `Pair` itself. The `Report`'s `scoredPairs` field is removed in favor of two methods: `getPair(file1, file2)` and `allPairs()`. 
- The `Pair` class no longer has a `fragments` public field, instead you need to call `buildFragments()` to create them. Other fragment-related methods like `squash()` are removed.

# Benchmarks

While solving performance issues, I've sun a benchmark using the `time ` program, recording wall time and reserved (RES) memory. The benchmark datasets were three medium exercises (390 files) and one large exercise (1162 files).

## Baseline

Before any changes. (commit 09d6416d9424fc0408a9329d25d3b74268729364)

| Exercise | Time    | Memory       |
|----------|---------|--------------|
| ISBN     | 14.57 s | 2 678 460 kB |
| Uitlezingen | 19.98 s | 3 176 008 kB |
| Autogram | error   | error        |

Error is "Maximum call stack size exceeded".

## Avoid splats (`...`) which cause the call stack error

| Exercise | Time | Memory |
|----------|------|--------|
| ISBN | 15.62 s | 2 758 740 kB |
| Uitlezingen | 19.38 s | 3 257 804 kB |
| Autogram | 16.58 s | 1 918 096 kB |

Error is gone, but memory usage is slightly increased.

## Make index from `SharedFingeprrints`

| Exercise    | Time    | Memory       |
|-------------|---------|--------------|
| ISBN        | 16.80 s | 2 673 532 kB |
| Uitlezingen | 20.29 s | 3 244 548 kB |
| Autogram    | 18.03 s | 1 845 864 kB |

Decrease in memory usage, but slower execution time.

## Calculate score without fragments

The most memory intensive step is building the `Fragments` for each `Pair` in the report. These fragments are used to calculate the score, however since #661 the fragmentsthemselves are not really needed anymore except when using the cli with the `--compare` flag.

This change calculates the scores (similarity, total overlap and longest overlap) without building the fragments. The most intensive step is calculating longest overlap, which is essentially searching for the longest common substring.

| Exercise | Time | Memory |
|----------|------|--------|
| ISBN | 5.66 s | 425 876 kB |
| Uitlezingen | 5.87 s | 283 336 kB |
| Autogram | 4.33 s | 243 560 kB |
| Plutokiller | 35.05 s | 915 788 kB |

The memory usage is reduced by an order of magnitude and Dolos also runs around 4 times faster.

## Store SharedFingerprints and Occurrences as sparse Array instead of Map

| Exercise | Time | Memory |
|----------|------|--------|
| ISBN | 5.11 s | 446 256 kB |
| Uitlezingen | 5.45 s | 304 552 kB |
| Autogram | 4.05 s | 273 184 kB |
| Plutokiller | 31.15 s | 1 026 208 kB |

Faster, but increased memory usage. Since memory is often more constrained than time, we will revert this change.

# Further possible improvements

## Tokenization

Profiling the large exercise gives the following flamegraph:
![image](https://user-images.githubusercontent.com/3226995/213725554-5c5110c4-fc05-4572-a071-048a454e7457.png)

There are three main components:
- Parsing and tokenizing files
- Building SharedFragments by calculating hashes and storing them in the index
- Calculating the pairs with their scores

The first item (parsing and tokenizing files) was not looked at in this optimization round and is a possible lead for further memory and speed improvements.
- The tokenization is a recursive function which can probably be sped up by walking the syntax tree in a more optimized manner
- We only need to calculate the AST in order to create hashed kgrams. The file content and AST are probably no longer needed once this has happened, but they are not GC'd because they are part of the `File` class. A lot of memory could be gained by allowing this data to be freed.

## Data serialization

We now (mis)use CSV files to pass data to the Web UI, but this is done in a less-than-optimal manner. For example: we write the full file content and the AST as fields in this CSV.

We should look into improving the data serialization by using another format (for example: SQLite).

## Web UI

This PR did not profile the Web UI.